Ajustando contextos e iniciando serviço de cadastros
