--- src/Modules/Cadastro/ControlePedidos.Cadastro.Application/Services/Cliente.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ControlePedidos.Cadastro.Application.Services
-{
-    public class ClienteService : IClienteValidationService
-    {
-        public bool ValidarCPF(string cpf)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool ValidarEmail(string email)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
