+++ src/Modules/Cadastro/ControlePedidos.Cadastro.Domain/Entities/Cliente.cs
@@ -8,12 +8,17 @@ namespace ControlePedidos.Cadastro.Domain.Entities
 {
     public class Cliente
     {
+        public Cliente()
+        {
+            
+        }
         public Cliente(string email, string cpf, string nome)
         {
             Email = email;
             CPF = cpf;
             Nome = nome;
         }
+        
         public string Email { get; set; }
         public string CPF { get; private set; }
         public string Nome { get; set; }
